mldsa: drop const generic in Operations trait
Eurydice can't handle associated types right now.
So this changes it such that the Operations trait is implemented
on top of a wrapper struct for the coefficients again.
This requires the implementations of the trait to point into the struct
again. But this is more self contained than propagating the generic type
for the trait.

I couldn't measure an impact on performance or stack size with this change