sha3: don't inline(always) sha3 API functions
Background:
Those API functions shouldn't be must inline in sha3. msvc is throwing
the sha3 functions away because they are KRML_MUSTINLINE (translated
from #[inline(always)]). I dropped those inlines.
We should double check performance. But for extractions that's
what's needed, and also in general what we should do.
If we need the inlines for performance, we need to expose the functions
separately to allow C to call them.